consolidate generation components
