+++ app/images/[thread_id]/page.tsx
@@ -1,35 +1,25 @@
 'use client'
 
-import { useState } from 'react'
-import * as Icons from '@phosphor-icons/react/dist/ssr'
 import { ScrollArea } from '@radix-ui/themes'
 import { usePathname } from 'next/navigation'
 
 import { useImagesQueryContext } from '@/app/images/ImagesQueryProvider'
-import { GenerationForm } from '@/components/generation/GenerationForm'
 import { IImageCard } from '@/components/images/IImageCard'
-import { IconButton } from '@/components/ui/Button'
 import { InfiniteScroll } from '@/components/ui/InfiniteScroll'
 import { Orbit } from '@/components/ui/Ldrs'
-import { SectionHeader } from '@/components/ui/Section'
-import { useThread, useThreadActions } from '@/lib/api'
 import { cn, twx } from '@/lib/utils'
 
 const ResultsGrid = twx.div`grid auto-rows-max grid-cols-2 md:grid-cols-3 gap-2 p-2 xl:grid-cols-4`
 
 export default function Page({ params }: { params: { thread_id: string } }) {
   const pathname = usePathname()
-  const thread = useThread(params.thread_id)
+
   const { imagesFeed } = useImagesQueryContext()
-
-  const actions = useThreadActions(thread?._id)
-
-  const [showGenerate, setShowGenerate] = useState(false)
 
   return (
     <>
       <div className="flex h-96 grow items-stretch divide-x overflow-hidden">
-        <div className={cn('flex grow flex-col overflow-hidden', showGenerate && 'hidden sm:flex')}>
+        <div className={cn('flex grow flex-col overflow-hidden')}>
           <ScrollArea scrollbars="vertical">
             <ResultsGrid className="mb-16 [&>div]:aspect-square">
               {imagesFeed.results.map((image) => (
@@ -65,27 +55,6 @@ export default function Page({ params }: { params: { thread_id: string } }) {
             </div>
           </ScrollArea>
         </div>
-
-        {showGenerate ? (
-          <aside className="flex w-full shrink-0 flex-col sm:w-80">
-            <SectionHeader className="px-2">Generate</SectionHeader>
-            <ScrollArea scrollbars="vertical" className="[&>div>div]:max-w-full">
-              <GenerationForm
-                onRun={actions.run}
-                loading={actions.state !== 'ready'}
-                storageKey={thread?.slug}
-              />
-            </ScrollArea>
-          </aside>
-        ) : null}
-
-        <IconButton
-          onClick={() => setShowGenerate(!showGenerate)}
-          aria-label={showGenerate ? 'Close' : 'Generate'}
-          className="absolute right-1 top-1"
-        >
-          {showGenerate ? <Icons.X /> : <Icons.Plus />}
-        </IconButton>
       </div>
     </>
   )
