--- app/images/page.tsx
@@ -1,11 +1,0 @@
-import { GenerationSection } from '@/components/generation/GenerationSection'
-import { Section } from '@/components/ui/Section'
-
-export default function Page() {
-  return (
-    <>
-      <GenerationSection />
-      <Section></Section>
-    </>
-  )
-}
