+++ components/generations/__fixtures__/cosmos.decorator.tsx
